requirements.txt
# ===================================
# IoT Sensor Platform - Dependencies
# ===================================

# Django Core
Django==5.0.1
django-cors-headers==4.3.1
python-decouple==3.8

# Django REST Framework
drf-spectacular==0.27.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg[binary]==3.2.13 

# Date and Time Utilities
python-dateutil==2.8.2
pytz==2024.1

# MQTT Support
paho-mqtt==1.6.1

# Development Tools
ipython==8.20.0
